🔨 feat: move oauth, auth to global from domain
