--- fixadate/src/main/java/com/fixadate/domain/auth/controller/AuthController.java
@@ -1,4 +1,0 @@
-package com.fixadate.domain.auth.controller;
-
-public class AuthController {
-}
