--- fixadate/src/main/java/com/fixadate/domain/auth/service/AuthService.java
@@ -1,4 +1,0 @@
-package com.fixadate.domain.auth.service;
-
-public class AuthService {
-}
