--- fixadate/src/main/java/com/fixadate/domain/jwt/controller/JwtController.java
@@ -1,4 +1,0 @@
-package com.fixadate.domain.jwt.controller;
-
-public class JwtController {
-}
