--- fixadate/src/main/java/com/fixadate/domain/jwt/exception/JwtException.java
@@ -1,4 +1,0 @@
-package com.fixadate.domain.jwt.exception;
-
-public class JwtException {
-}
