--- fixadate/src/main/java/com/fixadate/domain/jwt/filter/JwtRequestFilter.java
@@ -1,4 +1,0 @@
-package com.fixadate.domain.jwt.filter;
-
-public class JwtRequestFilter {
-}
