--- fixadate/src/main/java/com/fixadate/domain/jwt/service/JwtService.java
@@ -1,4 +1,0 @@
-package com.fixadate.domain.jwt.service;
-
-public class JwtService {
-}
