+++ fixadate/src/main/java/com/fixadate/global/auth/controller/AuthController.java
@@ -1,0 +1,4 @@
+package com.fixadate.global.auth.controller;
+
+public class AuthController {
+}
