+++ fixadate/src/main/java/com/fixadate/global/auth/service/AuthService.java
@@ -1,0 +1,4 @@
+package com.fixadate.global.auth.service;
+
+public class AuthService {
+}
