+++ fixadate/src/main/java/com/fixadate/global/jwt/controller/JwtController.java
@@ -1,0 +1,4 @@
+package com.fixadate.global.jwt.controller;
+
+public class JwtController {
+}
