+++ fixadate/src/main/java/com/fixadate/global/jwt/exception/JwtException.java
@@ -1,0 +1,4 @@
+package com.fixadate.global.jwt.exception;
+
+public class JwtException {
+}
