+++ fixadate/src/main/java/com/fixadate/global/jwt/filter/JwtExceptionFilter.java
@@ -1,0 +1,4 @@
+package com.fixadate.global.jwt.filter;
+
+public class JwtExceptionFilter {
+}
