+++ fixadate/src/main/java/com/fixadate/global/jwt/filter/JwtRequestFilter.java
@@ -1,0 +1,4 @@
+package com.fixadate.global.jwt.filter;
+
+public class JwtRequestFilter {
+}
