+++ fixadate/src/main/java/com/fixadate/global/jwt/service/JwtService.java
@@ -1,0 +1,4 @@
+package com.fixadate.global.jwt.service;
+
+public class JwtService {
+}
